added globals file, separated components into individual files, some other stuff
