+++ .gitignore
@@ -10,3 +10,4 @@ assets/**/*.mp4
 assets/**/*.obj
 assets/**/*.gltf
 dist
+archive
