+++ index.html
@@ -15,7 +15,7 @@
         <img id="360-test" src="assets/360-photo.jpg">
       </a-assets>
       <!-- scene--> 
-      <a-entity id="scene-template" template="src: templates/scene_home.html">
+      <a-entity id="scene-template" template="src:templates/scene_home.html">
       </a-entity> 
       <!-- managers--> 
       <a-entity nav-manager timeline-manager scene-manager test-manager></a-entity>
@@ -29,7 +29,7 @@
              fill="backwards" from="0.1 0.1 0.1" to="1 1 1"></a-animation>
         <a-animation begin="cursor-fusing" easing="ease-in" attribute="scale"
              fill="forwards" from="1 1 1" to="0.1 0.1 0.1"></a-animation>
-       </a-entity>
+       </a-entity>           
       </a-camera>
     </a-scene>
