--- js/main.json
@@ -1,1 +1,0 @@
-{"locations":{"origin":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ns-1":{"description":"Centerpoint of interior","coord":{"x":12,"y":0,"z":12}},"ns-2":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ns-3":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ns-5":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ns-6":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ew-1":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ew-2":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ew-3":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}},"ew-4":{"description":"Centerpoint of interior","coord":{"x":0,"y":0,"z":0}}},"dates":{"Mon Oct 31 2016 17:00:00 GMT-0700 (PDT)":[{"type":"model","title":"First forest model","description":"Initial 3D model","location":"origin","source":"16-11-temple.gltf"}],"Wed Nov 30 2016 16:00:00 GMT-0800 (PST)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"16-12-temple.gltf"}],"Sat Dec 31 2016 16:00:00 GMT-0800 (PST)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.gltf"}],"Fri Mar 31 2017 17:00:00 GMT-0700 (PDT)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.gltf"}],"Sun Apr 30 2017 17:00:00 GMT-0700 (PDT)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.gltf"}],"Wed May 31 2017 17:00:00 GMT-0700 (PDT)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.gltf"}],"Fri Jun 30 2017 17:00:00 GMT-0700 (PDT)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.gltf"}],"Mon Jul 31 2017 17:00:00 GMT-0700 (PDT)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"model/temple-blender-no-material.gltf"}],"Mon Aug 14 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Tue Aug 15 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Wed Aug 16 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Thu Aug 17 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Fri Aug 18 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Sat Aug 19 2017 17:00:00 GMT-0700 (PDT)":[{"type":"model","title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.gltf"},{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Sun Aug 20 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Mon Aug 21 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Tue Aug 22 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Wed Aug 23 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Thu Aug 24 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Fri Aug 25 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Sat Aug 26 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}],"Sun Aug 27 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"main01","source":"somesorce.jpg"}],"Sat Sep 02 2017 17:00:00 GMT-0700 (PDT)":[{"type":360,"title":"Design iteration model","description":"Initial 3D model","location":"origin","source":"somesorce.jpg"}]}}
